authorization implemented

+++ src/api/notes/handler.js
@@ -18,11 +18,15 @@ class NotesHandler {
             const {
                 title = 'untitled', body, tags
             } = request.payload;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
 
             const noteId = await this._service.addNote({
                 title,
                 body,
-                tags
+                tags,
+                owner: credentialId,
             });
 
             const response = h.response({
@@ -52,8 +56,11 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request) {
+        const {
+            id: credentialId
+        } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
         return {
             status: 'success',
             data: {
@@ -67,6 +74,11 @@ class NotesHandler {
             const {
                 id
             } = request.params;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
             const note = await this._service.getNoteById(id);
             return {
                 status: 'success',
@@ -102,6 +114,12 @@ class NotesHandler {
                 id
             } = request.params;
 
+            const {
+                id: credentialId
+            } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
+
             await this._service.editNoteById(id, request.payload);
 
             return {
@@ -133,6 +151,11 @@ class NotesHandler {
             const {
                 id
             } = request.params;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
             await this._service.deleteNoteById(id);
             return {
                 status: 'success',
